feat: Handles exception

+++ src/main/java/com/project/rest/usersRestApi/controller/UserController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.project.rest.usersRestApi.handler.BusinessException;
 import com.project.rest.usersRestApi.modal.User;
 import com.project.rest.usersRestApi.repository.UserRepository;
 
@@ -33,24 +35,54 @@ public class UserController {
 
 	@PostMapping
 	public User addUser(@RequestBody User user) {
+		validateUser(user);
+		if (repository.findByUserName(user.getUserName()).isPresent()) {
+			throw new BusinessException("Username already exists. Try with another username");
+		}
 		return repository.save(user);
 	}
 
 	@PutMapping
 	public User put(@RequestBody User user) {
-		if (user.getId() != null && user.getId() > 0)
+		validateUser(user);
+		if (user.getId() != null && user.getId() >= 0) {
+			User existingUser = repository.findById(user.getId())
+					.orElseThrow(() -> new BusinessException("User not found. Try with another id"));
+			if (!existingUser.getUserName().equals(user.getUserName())
+					&& repository.findByUserName(user.getUserName()).isPresent()) {
+				throw new BusinessException("Username already exists. Try with another username");
+			}
 			return repository.save(user);
-		return user;
+
+		} else {
+			throw new BusinessException("User not found. Try with another id");
+		}
 	}
 
 	@DeleteMapping
 	public String deleteUser(@RequestBody User user) {
-		if (user.getId() > 0) {
+		if (user.getId() != null && user.getId() >= 0) {
+			User existingUser = repository.findById(user.getId())
+					.orElseThrow(() -> new BusinessException("User not found. Try with another id"));
+
+			existingUser.getId();
 			repository.delete(user);
 			return "User deleted";
+		} else {
+			throw new BusinessException("User not found. Try with another id");
 		}
-		
-		return "User not found";
+	}
+
+	private void validateUser(User user) {
+		if (user.getName() == null || user.getName().isEmpty()) {
+			throw new BusinessException("Name is required");
+		}
+		if (user.getUserName() == null || user.getUserName().isEmpty()) {
+			throw new BusinessException("Username is required");
+		}
+		 if (user.getPassword() == null || user.getPassword().isEmpty()) {
+	            throw new BusinessException("Password is required");
+	        }
 	}
 
 }
